requirements.txt
asgiref==3.8.1
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
sqlparse==0.5.3
django-environ==0.12.0
python-dotenv==1.1.0
django-cors-headers==4.7.0
drf-spectacular==0.28.0
djoser==2.3.1
